requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
sqlalchemy>=2.0.30
psycopg2-binary>=2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart>=0.0.9
pydub>=0.25.1
librosa>=0.10.1
soundfile>=0.12.1
numpy>=1.24.0
slowapi==0.1.4

# ML — pinned for reproducible builds
torch==2.2.2
torchaudio==2.2.2
speechbrain>=1.0.0
huggingface-hub>=0.23.2,<1.0
scipy>=1.11.0
faiss-cpu>=1.7.4

gunicorn>=21.2.0
requests
python-dotenv
